+++ main.go
@@ -20,265 +20,39 @@ package main
 import (
 	"context"
 	"log"
-	"net/http"
 	"os"
-	"time"
 
 	"github.com/gin-gonic/gin"
-	"gopkg.in/mgo.v2/bson"
+	"github.com/m0rk0vka/gin-demo/handlers"
 
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/mongo/readpref"
 )
 
-var recipes []Recipe
-var ctx context.Context
-var err error
-var client *mongo.Client
-var collection *mongo.Collection
+var recipesHandler *handlers.RecipesHandler
 
 func init() {
-	//recipes = make([]Recipe, 0)
-	//file, _ := os.ReadFile("recipes.json")
-	//_ = json.Unmarshal(file, &recipes)
-
-	ctx = context.Background()
-	client, err = mongo.Connect(ctx, options.Client().ApplyURI(os.Getenv("MONGO_URI")))
-	if err = client.Ping(context.TODO(), readpref.Primary()); err != nil {
+	ctx := context.Background()
+	client, err := mongo.Connect(ctx, options.Client().ApplyURI(os.Getenv("MONGO_URI")))
+	if err != nil {
 		log.Fatal(err)
 	}
 
-	collection = client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-
+	if err = client.Ping(context.TODO(), readpref.Primary()); err != nil {
+		log.Fatal(err)
+	}
 	log.Println("Connected to MongoDB")
-
-	//var listOfRecipes []interface{}
-	//for _, recipe := range recipes {
-	//	listOfRecipes = append(listOfRecipes, recipe)
-	//}
-	//collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-	//insertManyResult, err := collection.InsertMany(ctx, listOfRecipes)
-	//if err != nil {
-	//	log.Fatal(err)
-	//}
-	//log.Println("Inserted recipes: ", len(insertManyResult.InsertedIDs))
+	collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
+	recipesHandler = handlers.NewRecipesHandler(ctx, collection)
 }
 
 func main() {
 	router := gin.Default()
-	router.POST("/recipes", NewRecipeHandler)
-	router.GET("/recipes", ListRecipesHandler)
-	router.PUT("/recipes/:id", UpdateRecipeHandler)
-	router.DELETE("/recipes/:id", DeleteRecipeHandler)
-	router.GET("/recipes/search", SearchRecipeHandler)
+	router.POST("/recipes", recipesHandler.NewRecipeHandler)
+	router.GET("/recipes", recipesHandler.ListRecipesHandler)
+	router.PUT("/recipes/:id", recipesHandler.UpdateRecipeHandler)
+	router.DELETE("/recipes/:id", recipesHandler.DeleteRecipeHandler)
+	router.GET("/recipes/search", recipesHandler.SearchRecipeHandler)
 	router.Run()
-}
-
-// swagger:parameters recipes newRecipe
-type Recipe struct {
-	// swagger:ignore
-	ID           primitive.ObjectID `json:"id" bson:"_id"`
-	Name         string             `json:"name" bson:"name"`
-	Tags         []string           `json:"tags" bson:"tags"`
-	Ingredients  []string           `json:"ingredients" bson:"ingredients"`
-	Instructions []string           `json:"instructions" bson:"instructions"`
-	PublishedAt  time.Time          `json:"publishedAt" bson:"publishedAt"`
-}
-
-// swagger:operation POST /recipes recipes createRecipe
-// Create a new recipe
-// ---
-// produces:
-// - application/json
-// responses:
-//
-//	   '200':
-//	     description: successful operation
-//	   '400':
-//			  description: invalid input
-func NewRecipeHandler(c *gin.Context) {
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	recipe.ID = primitive.NewObjectID()
-	recipe.PublishedAt = time.Now()
-	_, err = collection.InsertOne(ctx, recipe)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError,
-			gin.H{"error": "Error while inserting a new recipe"})
-		return
-	}
-
-	c.JSON(http.StatusOK, recipe)
-}
-
-// swagger:operation GET /recipes recipes listRecipes
-// Returns list of recipes
-// ---
-// produces:
-// - application/json
-// responses:
-//
-//	 '200':
-//			description: Successful operation
-//
-//	 '500':
-//
-//		description: Error while executing find comand to collection
-func ListRecipesHandler(c *gin.Context) {
-	cur, err := collection.Find(ctx, bson.M{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError,
-			gin.H{"error": err.Error()})
-		return
-	}
-	defer cur.Close(ctx)
-
-	recipes := make([]Recipe, 0)
-	for cur.Next(ctx) {
-		var recipe Recipe
-		cur.Decode(&recipe)
-		recipes = append(recipes, recipe)
-	}
-
-	c.JSON(http.StatusOK, recipes)
-}
-
-// swagger:operation PUT /recipes/{id} recipes updateRecipe
-// Update an existing recipe
-// ---
-// parameters:
-//   - name: id
-//     in: path
-//     description: id of the recipe
-//     required: true
-//     type: string
-//
-// produces:
-// - application/json
-// responses:
-//
-//	   '200':
-//	     description: successful operation
-//	   '400':
-//			  description: invalid input
-//	   '500':
-//			  description: error while update db
-func UpdateRecipeHandler(c *gin.Context) {
-	id := c.Params.ByName("id")
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	objectId, _ := primitive.ObjectIDFromHex(id)
-	filter := bson.M{"_id": objectId}
-	update := bson.M{"$set": bson.M{
-		"name":         recipe.Name,
-		"instructions": recipe.Instructions,
-		"ingredients":  recipe.Ingredients,
-		"tags":         recipe.Tags,
-	}}
-	opts := options.Update().SetUpsert(true)
-	_, err := collection.UpdateOne(ctx, filter, update, opts)
-
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"meddage": "Recipe has been updated"})
-}
-
-// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
-// Delete an existing recipe
-// ---
-// parameters:
-//   - name: id
-//     in: path
-//     description: id of the recipe
-//     required: true
-//     type: string
-//
-// produces:
-// - application/json
-// responses:
-//
-//	   '200':
-//	     description: successful operation
-//	   '500':
-//			  description: error while delete recipe
-func DeleteRecipeHandler(c *gin.Context) {
-	id := c.Params.ByName("id")
-	objectId, _ := primitive.ObjectIDFromHex(id)
-	filter := bson.M{"_id": objectId}
-	_, err = collection.DeleteOne(ctx, filter)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Recipe has been deleted",
-	})
-}
-
-// swagger:operation GET /recipes/search recipes findRecipe
-// Search reipes based on tag
-// ---
-// parameters:
-//   - name: tag
-//     in: query
-//     description: recipe tag
-//     required: true
-//     type: string
-//
-// produces:
-// - application/json
-// responses:
-//
-//	 '200':
-//		  description: successful operation
-//	 '500':
-//			  description: error while executing find command on collection
-func SearchRecipeHandler(c *gin.Context) {
-	tag := c.Query("tag")
-	listOfRecipes := make([]Recipe, 0)
-
-	cur, err := collection.Find(ctx, bson.M{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-	}
-	defer cur.Close(ctx)
-
-	for cur.Next(ctx) {
-		var recipe Recipe
-		cur.Decode(&recipe)
-		find := false
-		for _, t := range recipe.Tags {
-			if t == tag {
-				find = true
-			}
-		}
-		if find {
-			listOfRecipes = append(listOfRecipes, recipe)
-		}
-	}
-
-	c.JSON(http.StatusOK, listOfRecipes)
 }
